# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: MindSafe Mental Health Platform API
  version: 1.0.0
  description: |
    Privacy-first AI mental health platform. All endpoints under `/api` are served by the
    API Gateway (Express) which proxies ML operations to internal Python microservices.

    ## Authentication
    Most endpoints require a Bearer token in the `Authorization` header.
    Obtain tokens via `/api/login` or `/api/auth/google`. Access tokens expire in 15 minutes;
    use `/api/refresh-token` to get a new pair.

    ## Error Format
    All errors return `{ "message": "..." }` with an appropriate HTTP status code.
  contact:
    name: MindSafe Team
  license:
    name: Private
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://mindsafe-api.onrender.com
    description: Free-tier production (Render)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `/api/login`. Expires in 15 minutes.
  schemas:
...
    CrisisDetectionRequest:
      type: object
      required: [user_id, indicators]
      properties:
        user_id:
          type: string
        indicators:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
              severity:
                type: string
                enum: [low, medium, high, critical]
        context:
          type: string
        recent_mood_trend:
          type: string
    CrisisResponsePlan:
      type: object
      properties:
        crisis_id:
          type: string
          format: uuid
        user_id:
          type: string
        detected_level:
          type: string
          enum: [LOW, MEDIUM, HIGH, CRITICAL]
        actions_taken:
          type: array
          items:
            type: string
        resources_provided:
          type: array
          items:
            type: object
        notifications_sent:
          type: array
          items:
            type: string
        escalation_status:
          type: string
        follow_up_scheduled:
          type: string
          format: date-time
        timestamp:
          type: string
          format: date-time

    # ── Profile ──
...
paths:
...
  /detect:
    post:
      tags: [Crisis Detection]
      summary: Detect crisis from indicators
      operationId: detectCrisis
      description: "**Internal service** (port 8003)."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrisisDetectionRequest"
      responses:
        "200":
          description: Crisis response plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrisisResponsePlan"
  /resources/{region}:
    get:
      tags: [Crisis Detection]
      summary: Get emergency resources by region
      operationId: getResources
      description: "**Internal service** (port 8003)."
      parameters:
        - name: region
          in: path
          required: true
          schema:
            type: string
            example: IN
      responses:
        "200":
          description: Regional emergency resources
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  region:
                    type: string
                  resources:
                    type: array
                    items:
                      type: object